♻️ Flatten main setting page
